+++ sdk/keyvault/keyvault-secrets/src/secretsModels.ts
@@ -31,7 +31,7 @@ export interface SecretClientInterface {
    * The getDeletedSecret method returns the specified deleted secret along with its properties.
    * This operation requires the secrets/get permission.
    */
-  getDeletedSecret(secretName: string, options?: coreHttp.OperationOptions): Promise<DeletedSecret>;
+  getDeletedSecret(secretName: string, options?: DeleteSecretOptions): Promise<DeletedSecret>;
 }
 
 /**
@@ -192,6 +192,20 @@ export interface SecretPollerOptions extends coreHttp.OperationOptions {
 
 /**
  * @interface
+ * An interface representing the optional parameters that can be
+ * passed to {@link beginDeleteSecret}
+ */
+export interface BeginDeleteSecretOptions extends SecretPollerOptions {}
+
+/**
+ * @interface
+ * An interface representing the optional parameters that can be
+ * passed to {@link beginRecoverDeletedSecret}
+ */
+export interface BeginRecoverDeletedSecretOptions extends SecretPollerOptions {}
+
+/**
+ * @interface
  * An interface representing the optional parameters that can be passed to {@link setSecret}.
  */
 export interface SetSecretOptions extends coreHttp.OperationOptions {
@@ -289,7 +303,27 @@ export interface RestoreSecretBackupOptions extends coreHttp.OperationOptions {}
 export interface RecoverDeletedSecretOptions extends coreHttp.OperationOptions {}
 
 /**
+ * @internal
+ * @ignore
  * @interface
- * An interface representing the optional parameters that can be passed to {@link listPropertiesOfSecretVersions}, {@link listPropertiesOfSecrets} and {@link listDeletedSecrets}.
+ * An interface representing the optional parameters that can be passed to {@link deleteSecret}.
  */
-export interface ListOperationOptions extends coreHttp.OperationOptions {}
+export interface DeleteSecretOptions extends coreHttp.OperationOptions {}
+
+/**
+ * @interface
+ * An interface representing the optional parameters that can be passed to {@link listPropertiesOfSecretVersions}.
+ */
+export interface ListPropertiesOfSecretVersionsOptions extends coreHttp.OperationOptions {}
+
+/**
+ * @interface
+ * An interface representing the optional parameters that can be passed to {@link listPropertiesOfSecrets}.
+ */
+export interface ListPropertiesOfSecretsOptions extends coreHttp.OperationOptions {}
+
+/**
+ * @interface
+ * An interface representing the optional parameters that can be passed to {@link listDeletedSecrets}.
+ */
+export interface ListDeletedSecretsOptions extends coreHttp.OperationOptions {}